requirements.txt
python-telegram-bot==13.13

# for .env file support
python-dotenv